add: Chat service connect external Social service, Social service CORS config; update: Chat service validate;

+++ chat-service/src/main/java/com/owl/chat_service/application/service/admin/chat_member/ControlChatMemberAdminSerivces.java
@@ -9,7 +9,9 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import com.owl.chat_service.application.service.admin.chat.GetChatAdminServices;
+import com.owl.chat_service.domain.chat.service.ChatMemberServices;
 import com.owl.chat_service.domain.chat.validate.ChatMemberValidate;
+import com.owl.chat_service.external_service.client.BlockUserServiceApiClient;
 import com.owl.chat_service.external_service.client.UserServiceApiClient;
 import com.owl.chat_service.persistence.mongodb.document.Chat;
 import com.owl.chat_service.persistence.mongodb.document.ChatMember;
@@ -21,16 +23,19 @@ import com.owl.chat_service.presentation.dto.admin.ChatMemberAdminRequest;
 @Service
 @Transactional
 public class ControlChatMemberAdminSerivces {
+
+    private final BlockUserServiceApiClient blockUserServiceApiClient;
     private final ChatMemberRepository chatMemberRepository;
     private final GetChatMemberAdminServices getChatMemberAdminServices;
     private final GetChatAdminServices getChatAdminServices;
     private final UserServiceApiClient userServiceApiClient;
 
-    public ControlChatMemberAdminSerivces(ChatMemberRepository chatMemberRepository, GetChatMemberAdminServices getChatMemberAdminServices, GetChatAdminServices getChatAdminServices, UserServiceApiClient userServiceApiClient) {
+    public ControlChatMemberAdminSerivces(ChatMemberRepository chatMemberRepository, GetChatMemberAdminServices getChatMemberAdminServices, GetChatAdminServices getChatAdminServices, UserServiceApiClient userServiceApiClient, BlockUserServiceApiClient blockUserServiceApiClient) {
         this.chatMemberRepository = chatMemberRepository;
         this.getChatMemberAdminServices = getChatMemberAdminServices;
         this.getChatAdminServices = getChatAdminServices;
-        this.userServiceApiClient = userServiceApiClient;}
+        this.userServiceApiClient = userServiceApiClient;
+        this.blockUserServiceApiClient = blockUserServiceApiClient;}
 
     public ChatMember addNewChatMember(ChatMemberAdminRequest chatMemberRequest) {
         ChatMember newChatMember = new ChatMember();
@@ -58,7 +63,7 @@ public class ControlChatMemberAdminSerivces {
 
         if (chatMemberRequest.inviterId != null && !chatMemberRequest.inviterId.isBlank()) {
             if (getChatMemberAdminServices.getChatMemberByChatIdAndMemberId(chatMemberRequest.chatId, chatMemberRequest.inviterId) == null)
-                throw new SecurityException("Inviter " + chatMemberRequest.inviterId + " does not have permission to access this chat");
+                throw new SecurityException("Inviter does not have permission to access this chat");
         }
 
         int numberOfMember = getChatMemberAdminServices.getChatMembersByChatId(chat.getId(), -1, 0, true).size();
@@ -85,6 +90,15 @@ public class ControlChatMemberAdminSerivces {
 
         // if (!ChatMemberValidate.validateNickname(chatMemberRequest.nickname))
         //     throw new IllegalArgumentException("Invalid nickname");
+
+        try {
+            if (blockUserServiceApiClient.getUserBlockUser(chatMemberRequest.memberId, chatMemberRequest.inviterId) != null) {
+                throw new IllegalArgumentException("Inviter have been blocked by user");
+            }
+        }
+        catch (Exception e) {
+            throw new IllegalArgumentException(e.getMessage());
+        }
 
         newChatMember.setId(UUID.randomUUID().toString());
         newChatMember.setMemberId(chatMemberRequest.memberId);
@@ -205,6 +219,9 @@ public class ControlChatMemberAdminSerivces {
 
         if (existingChatMember == null) 
             throw new IllegalArgumentException("Chat member does not exists");
+
+        if (ChatMemberServices.compareRole(existingChatMember.getRole(), ChatMemberRole.OWNER) == 0)
+            throw new IllegalArgumentException("Chat owner cannot leave chat");
 
         chatMemberRepository.deleteById(Objects.requireNonNull(existingChatMember.getId(), "Delete chat member id is null"));
     }
